Refactor and tests
Adding a User aggregate root.
Adding a TaskCollection which is basically a collection of events the
User AR adds to which all gets dispatched.
Adding a UesrRepository in order to test out a different kind of
repository pattern.
Updating the UserListener class to act on the User AR which then
dispatches events to get acted upon by reactors, but for right now, I am
using the same UserListener.
Adding in a number of tests to make sure the basic logic works.