CircleCI
Adding circleci for testing which uses the phusion/baseimage:latest
since the Phalcon framework is an extension and needs to be installed.
This is the only image I've been able to get the framework installed on
so I'll just stick with it for now. Most of the run commands come from
the repository exilesprx/phaldock that I built.
Adding the README and LICENSE files